vertical quick action layout
